requirements.txt
pyusb
ttkthemes
pillow
pyserial
adb-shell
